refactor: make both original graphql/execute and Rx graphql/execute be testable via setting a ENV variable `RUN_MODE`
