+++ src/graphql.js
@@ -11,10 +11,30 @@
 import { Source } from './language/source';
 import { parse } from './language/parser';
 import { validate } from './validation/validate';
-import { execute } from './execution/execute';
 import type { GraphQLSchema } from './type/schema';
 import type { ExecutionResult } from './execution/execute';
 
+import { execute } from './execution/execute-origin';
+import { executeRx } from './execution/execute-rx';
+import { Observable } from 'rxjs/Rx';
+
+
+export function graphql(
+  schema: GraphQLSchema,
+  requestString: string,
+  rootValue?: mixed,
+  contextValue?: mixed,
+  variableValues?: ?{[key: string]: mixed},
+  operationName?: ?string
+): Promise<ExecutionResult> {
+  const mode = process.env.RUN_MODE;
+  if (mode === 'Rx') {
+    return graphqlRx(schema, requestString, rootValue,
+      contextValue, variableValues, operationName).toPromise();
+  }
+  return graphqlOri(schema, requestString, rootValue,
+    contextValue, variableValues, operationName);
+}
 
 /**
  * This is the primary entry point function for fulfilling GraphQL operations
@@ -40,7 +60,7 @@ import type { ExecutionResult } from './execution/execute';
  *    possible operations. Can be omitted if requestString contains only
  *    one operation.
  */
-export function graphql(
+export function graphqlOri(
   schema: GraphQLSchema,
   requestString: string,
   rootValue?: mixed,
@@ -68,5 +88,38 @@ export function graphql(
     }
   }).then(undefined, error => {
     return { errors: [ error ] };
+  });
+}
+
+export function graphqlRx(
+  schema: GraphQLSchema,
+  requestString: string,
+  rootValue?: mixed,
+  contextValue?: mixed,
+  variableValues?: ?{[key: string]: mixed},
+  operationName?: ?string
+): rxjs$Observable<ExecutionResult> {
+  return Observable.defer(() => {
+    try {
+      const source = new Source(requestString || '', 'GraphQL request');
+      const documentAST = parse(source);
+      const validationErrors = validate(schema, documentAST);
+      if (validationErrors.length > 0) {
+        return Observable.of({ errors: validationErrors });
+      }
+
+      return executeRx(
+        schema,
+        documentAST,
+        rootValue,
+        contextValue,
+        variableValues,
+        operationName
+      ).catch(error => {
+        return Observable.of({ errors: [ error ] });
+      });
+    } catch (error) {
+      return Observable.of({ errors: [ error ] });
+    }
   });
 }
